backend/app/services/data_collector.py
# backend/app/services/data_collector.py
import requests
import asyncio
from bs4 import BeautifulSoup
from datetime import datetime
import json


class PatentCollector:
    def __init__(self):
        self.base_url = "https://patents.google.com"

    async def search_patents(self, query: str, limit: int = 100):
        """Collect patents from Google Patents API"""
        # Mock data for demo - replace with actual API calls
        mock_patents = []
        for i in range(limit):
            patent = {
                "id": f"US{10000000 + i}",
                "title": f"{query} Patent {i+1}",
                "abstract": f"This patent describes innovative {query} technology...",
                "inventors": ["John Doe", "Jane Smith"],
                "assignee": f"Company {i % 10}",
                "filing_date": datetime.now(),
                "publication_date": datetime.now(),
                "classification": ["H01L", "G06N"],
                "country": "US",
            }
            mock_patents.append(patent)
        return mock_patents


class ResearchCollector:
    def __init__(self):
        self.base_url = "https://arxiv.org"

    async def search_papers(self, query: str, limit: int = 50):
        """Collect research papers from arXiv API"""
        # Mock data for demo
        mock_papers = []
        for i in range(limit):
            paper = {
                "id": f"arxiv:{2024}.{i:04d}",
                "title": f"Advances in {query}: A Comprehensive Study {i+1}",
                "abstract": f"This paper presents novel approaches to {query}...",
                "authors": ["Dr. Alice Johnson", "Prof. Bob Wilson"],
                "journal": "Nature Technology",
                "publication_date": datetime.now(),
                "keywords": [query, "artificial intelligence", "machine learning"],
                "citations": (i * 10) % 100,
            }
            mock_papers.append(paper)
        return mock_papers
